test(extension): fixes on PR
