+++ packages/e2e-tests/src/features/analytics/AnalyticsForgotPassword.feature
@@ -45,7 +45,7 @@ Feature: Analytics - Forgot Password
     And I close all remaining tabs except current one
     And I set up request interception for posthog analytics request(s)
     Then "Wallet setup" page is displayed
-    When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
+    # When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "unlock wallet | forgot password? | set up your password | next | click"
     And I click "Next" button during wallet setup
