+++ packages/e2e-tests/src/features/analytics/AnalyticsOnboardingEvents.feature
@@ -6,13 +6,13 @@ Feature: Analytics - Posthog - Onboarding - Extended View
   Scenario Outline: Analytics - Posthog events are enabled or disabled based on decision <enable_analytics> on Analytics page
     Given I set up request interception for posthog analytics request(s)
     When I click "Create" button on wallet setup page
-    When I accept "T&C" checkbox
+    # When I accept "T&C" checkbox
     And I click "Next" button during wallet setup
 #    When "Help us improve your experience" page is displayed
     And I click "<enable_analytics>" button on Analytics page
     And I enter wallet name: "wallet"
     And I click "Next" button during wallet setup
-    When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
+    # When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
     And I click "Next" button during wallet setup
 #    When "Mnemonic info" page is displayed
     And I validate that <number_of_events> analytics event(s) have been sent
@@ -28,7 +28,7 @@ Feature: Analytics - Posthog - Onboarding - Extended View
     Given I set up request interception for posthog analytics request(s)
     When I click "Restore" button on wallet setup page
     When I click "OK" button on "Restoring a multi-address wallet?" modal
-    When I accept "T&C" checkbox
+    # When I accept "T&C" checkbox
     And I click "Next" button during wallet setup
 #    When "Help us improve your experience" page is displayed
     And I click "Agree" button on Analytics page
@@ -39,17 +39,17 @@ Feature: Analytics - Posthog - Onboarding - Extended View
     And I enter wallet name: "wallet"
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "onboarding | restore wallet | wallet name | next | click"
-    When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
+    # When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "onboarding | restore wallet | wallet password | next | click"
 #    When "Recovery phrase length page" is displayed and 24 words checkbox is checked
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "onboarding | restore wallet | recovery phrase length | next | click"
-    When I pass "Mnemonic verification" page with words 8 of 24
+    # When I pass "Mnemonic verification" page with words 8 of 24
     Then I validate latest analytics single event "onboarding | restore wallet | enter passphrase #01 | next | click"
-    When I pass "Mnemonic verification" page with words 16 of 24
+    # When I pass "Mnemonic verification" page with words 16 of 24
     Then I validate latest analytics single event "onboarding | restore wallet | enter passphrase #09 | next | click"
-    When I pass "Mnemonic verification" page with words 24 of 24
+    # When I pass "Mnemonic verification" page with words 24 of 24
     Then I validate latest analytics single event "onboarding | restore wallet | enter passphrase #17 | next | click"
     When I click "Go to my wallet" button on "All done" page
     And I validate latest analytics multiple events:
@@ -63,7 +63,7 @@ Feature: Analytics - Posthog - Onboarding - Extended View
   Scenario: Analytics - Onboarding new wallet events
     Given I set up request interception for posthog analytics request(s)
     When I click "Create" button on wallet setup page
-    When I accept "T&C" checkbox
+   #  When I accept "T&C" checkbox
     And I click "Next" button during wallet setup
 #    When "Help us improve your experience" page is displayed
     And I click "Agree" button on Analytics page
@@ -74,23 +74,23 @@ Feature: Analytics - Posthog - Onboarding - Extended View
     And I enter wallet name: "wallet"
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "onboarding | new wallet | wallet name | next | click"
-    When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
+    # When I enter password: "N_8J@bne87A" and password confirmation: "N_8J@bne87A"
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "onboarding | new wallet | wallet password | next | click"
 #    When "Mnemonic info" page is displayed
     And I click "Next" button during wallet setup
     Then I validate latest analytics single event "onboarding | new wallet | passphrase intro | next | click"
-    When I pass "Mnemonic writedown" page with words 8 of 24
+    # When I pass "Mnemonic writedown" page with words 8 of 24
     Then I validate latest analytics single event "onboarding | new wallet | write passphrase #01 | next | click"
-    When I pass "Mnemonic writedown" page with words 16 of 24
+    # When I pass "Mnemonic writedown" page with words 16 of 24
     Then I validate latest analytics single event "onboarding | new wallet | write passphrase #09 | next | click"
-    When I pass "Mnemonic writedown" page with words 24 of 24
+    # When I pass "Mnemonic writedown" page with words 24 of 24
     Then I validate latest analytics single event "onboarding | new wallet | write passphrase #17 | next | click"
-    When I pass "Mnemonic verification" page with words 8 of 24
+    # When I pass "Mnemonic verification" page with words 8 of 24
     Then I validate latest analytics single event "onboarding | new wallet | enter passphrase #01 | next | click"
-    When I pass "Mnemonic verification" page with words 16 of 24
+    # When I pass "Mnemonic verification" page with words 16 of 24
     Then I validate latest analytics single event "onboarding | new wallet | enter passphrase #09 | next | click"
-    When I pass "Mnemonic verification" page with words 24 of 24
+    # When I pass "Mnemonic verification" page with words 24 of 24
     Then I validate latest analytics single event "onboarding | new wallet | enter passphrase #17 | next | click"
     When I click "Go to my wallet" button on "All done" page
     And I see LW homepage
