+++ packages/e2e-tests/src/steps/forgotPasswordSteps.ts
@@ -2,14 +2,10 @@ import { Then, When } from '@cucumber/cucumber';
 import WalletUnlockPage from '../elements/walletUnlockPage';
 import ForgotPasswordModalAssert from '../assert/forgotPasswordModalAssert';
 import ForgotPasswordModal from '../elements/ForgotPasswordModal';
-import OnboardingPageObject from '../pageobject/onboardingPageObject';
-import WalletSetupPage from '../elements/onboarding/walletSetupPage';
 import LocalStorageAssert from '../assert/localStorageAssert';
 import BackgroundStorageAssert from '../assert/backgroundStorageAssert';
 import extendedView from '../page/extendedView';
 import { browser } from '@wdio/globals';
-
-const validPassword = 'N_8J@bne87A';
 
 When(/^I click on "Forgot password\?" button on unlock screen$/, async () => {
   await WalletUnlockPage.forgotPassword.waitForClickable();
@@ -44,20 +40,6 @@ Then(/^I switch to tab with restore wallet process$/, async () => {
       await browser.pause(1000);
     }
     retries--;
-  }
-});
-
-Then(/^I am on (.*) page of restoration flow$/, async (expectedPage: string) => {
-  switch (expectedPage) {
-    case 'password':
-      // nothing to do as user lands on Password Page by default
-      break;
-    case 'mnemonic verification':
-      await OnboardingPageObject.fillPasswordPage(validPassword, validPassword);
-      await WalletSetupPage.nextButton.click();
-      break;
-    default:
-      throw new Error(`Unknown page: ${expectedPage}`);
   }
 });
